Module ic_p2p::p2p

The P2P module exposes the peer-to-peer functionality.

Specifically, it constructs all the artifact pools and the Consensus/P2P time source.

Structs

P2PService
A helper service to run the P2P component.

P2P
The P2P struct, which encapsulates all relevant components including gossip and event handler control.

TestArtifact
The artifact struct used by the testing framework.

Enums

P2PStateSyncClient
The P2P state sync client.

Type Definitions

TestArtifactAttribute
The attribute of the artifact used by the testing framework.

TestArtifactId
The artifact ID used by the testing framework.

TestArtifactMessage
The artifact message used by the testing framework.